move peer list into domain
