+++ crates/domain/src/models/mod.rs
@@ -1,5 +1,7 @@
 pub mod block_index;
 pub mod block_tree;
+pub mod peer_list;
 
 pub use block_index::*;
 pub use block_tree::*;
+pub use peer_list::*;
