+++ crates/p2p/src/peer_list.rs
@@ -1,31 +1,27 @@
 use crate::types::GossipDataRequest;
 use crate::GossipClient;
 use actix::prelude::*;
-use alloy_core::primitives::B256;
 use irys_actors::reth_service::RethServiceActor;
 use irys_api_client::{ApiClient, IrysApiClient};
+use irys_database::insert_peer_list_item;
 use irys_database::reth_db::{Database as _, DatabaseError};
-use irys_database::tables::PeerListItems;
-use irys_database::{insert_peer_list_item, walk_all};
+use irys_domain::{
+    PeerListDataError, PeerListDataMessage, PeerListGuard, ScoreDecreaseReason, ScoreIncreaseReason,
+};
 use irys_types::{
-    build_user_agent, Address, BlockHash, Config, DatabaseProvider, PeerAddress, PeerListItem,
-    PeerResponse, RejectedResponse, RethPeerInfo, VersionRequest,
+    build_user_agent, Address, Config, DatabaseProvider, PeerAddress, PeerListItem, PeerResponse,
+    RejectedResponse, RethPeerInfo, VersionRequest,
 };
 use rand::prelude::SliceRandom as _;
 use std::collections::{HashMap, HashSet};
-use std::net::{IpAddr, SocketAddr};
-use std::sync::{Arc, RwLock};
+use std::net::SocketAddr;
 use std::time::Duration;
-use thiserror::Error;
-use tokio::sync::mpsc::error::SendError;
-use tokio::sync::mpsc::{UnboundedReceiver, UnboundedSender};
+use tokio::sync::mpsc::UnboundedReceiver;
 use tracing::{debug, error, info, warn};
 
 const FLUSH_INTERVAL: Duration = Duration::from_secs(5);
 const INACTIVE_PEERS_HEALTH_CHECK_INTERVAL: Duration = Duration::from_secs(10);
 const PEER_HANDSHAKE_RETRY_INTERVAL: Duration = Duration::from_secs(5);
-pub(crate) const MILLISECONDS_IN_SECOND: u64 = 1000;
-pub(crate) const HANDSHAKE_COOLDOWN: u64 = MILLISECONDS_IN_SECOND * 5;
 
 async fn send_message_and_print_error<T, A, R>(message: T, address: Addr<A>)
 where
@@ -40,416 +36,6 @@ where
                 "Failed to send message to peer service: {:?}",
                 mailbox_error
             );
-        }
-    }
-}
-
-#[derive(Debug, Error, Clone)]
-pub enum PeerListFacadeError {
-    #[error("Internal peer list service error: {0:?}")]
-    InternalError(String),
-    #[error("{0:?}")]
-    ServiceError(PeerListServiceError),
-}
-
-impl From<PeerListServiceError> for PeerListFacadeError {
-    fn from(err: PeerListServiceError) -> Self {
-        Self::ServiceError(err)
-    }
-}
-
-impl From<MailboxError> for PeerListFacadeError {
-    fn from(err: MailboxError) -> Self {
-        Self::InternalError(format!("{:?}", err))
-    }
-}
-#[derive(Debug, Clone)]
-pub struct PeerListDataInner {
-    gossip_addr_to_mining_addr_map: HashMap<IpAddr, Address>,
-    api_addr_to_mining_addr_map: HashMap<SocketAddr, Address>,
-    peer_list_cache: HashMap<Address, PeerListItem>,
-    known_peers_cache: HashSet<PeerAddress>,
-    trusted_peers_api_addresses: HashSet<SocketAddr>,
-    peer_list_service_sender: UnboundedSender<PeerListDataMessage>,
-}
-
-// TODO: remove default, right now it's there just to make service into a ssytem service
-impl Default for PeerListDataInner {
-    fn default() -> Self {
-        panic!("PeerListData should be initialized with new() method, not default()");
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct PeerListGuard(Arc<RwLock<PeerListDataInner>>);
-
-impl Default for PeerListGuard {
-    fn default() -> Self {
-        panic!("PeerListGuard should be initialized with new() method, not default()");
-    }
-}
-
-impl PeerListGuard {
-    pub fn new(
-        config: &Config,
-        db: &DatabaseProvider,
-        peer_service_sender: UnboundedSender<PeerListDataMessage>,
-    ) -> Result<Self, PeerListDataError> {
-        let inner = PeerListDataInner::new(config, db, peer_service_sender)?;
-        Ok(Self(Arc::new(RwLock::new(inner))))
-    }
-
-    pub fn add_or_update_peer(&self, mining_addr: Address, peer: PeerListItem) {
-        let mut inner = self.0.write().expect("PeerListDataInner lock poisoned");
-        inner.add_or_update_peer(mining_addr, peer);
-    }
-
-    pub fn increase_peer_score(&self, mining_addr: &Address, reason: ScoreIncreaseReason) {
-        let mut inner = self.0.write().expect("PeerListDataInner lock poisoned");
-        inner.increase_score(mining_addr, reason);
-    }
-
-    pub fn decrease_peer_score(&self, mining_addr: &Address, reason: ScoreDecreaseReason) {
-        let mut inner = self.0.write().expect("PeerListDataInner lock poisoned");
-        inner.decrease_peer_score(mining_addr, reason);
-    }
-
-    pub fn all_known_peers(&self) -> Vec<PeerAddress> {
-        self.read().known_peers_cache.iter().copied().collect()
-    }
-
-    pub async fn wait_for_active_peers(&self) {
-        loop {
-            let active_peers_count = {
-                let bindings = self.read();
-                bindings
-                    .peer_list_cache
-                    .values()
-                    .filter(|peer| peer.reputation_score.is_active() && peer.is_online)
-                    .count()
-            };
-
-            if active_peers_count > 0 {
-                return;
-            }
-
-            // Check for active peers every second
-            tokio::time::sleep(Duration::from_secs(1)).await;
-        }
-    }
-
-    pub fn trusted_peers(&self) -> Vec<(Address, PeerListItem)> {
-        let guard = self.read();
-
-        let mut peers: Vec<(Address, PeerListItem)> = guard
-            .peer_list_cache
-            .iter()
-            .map(|(key, value)| (*key, value.clone()))
-            .collect();
-
-        peers.retain(|(_miner_address, peer)| {
-            guard
-                .trusted_peers_api_addresses
-                .contains(&peer.address.api)
-        });
-
-        peers.sort_by_key(|(_address, peer)| peer.reputation_score.get());
-        peers.reverse();
-
-        peers
-    }
-
-    pub fn top_active_peers(
-        &self,
-        limit: Option<usize>,
-        exclude_peers: Option<HashSet<Address>>,
-    ) -> Vec<(Address, PeerListItem)> {
-        let guard = self.read();
-
-        let mut peers: Vec<(Address, PeerListItem)> = guard
-            .peer_list_cache
-            .iter()
-            .map(|(key, value)| (*key, value.clone()))
-            .collect();
-
-        peers.retain(|(miner_address, peer)| {
-            let exclude = if let Some(exclude_peers) = &exclude_peers {
-                exclude_peers.contains(miner_address)
-            } else {
-                false
-            };
-            !exclude && peer.reputation_score.is_active() && peer.is_online
-        });
-
-        peers.sort_by_key(|(_address, peer)| peer.reputation_score.get());
-        peers.reverse();
-
-        if let Some(truncate) = limit {
-            peers.truncate(truncate);
-        }
-
-        peers
-    }
-
-    pub fn peer_by_gossip_address(&self, address: SocketAddr) -> Option<PeerListItem> {
-        let binding = self.read();
-        let mining_address = binding
-            .gossip_addr_to_mining_addr_map
-            .get(&address.ip())
-            .copied()?;
-        binding.peer_list_cache.get(&mining_address).cloned()
-    }
-
-    pub fn peer_by_mining_address(&self, mining_address: &Address) -> Option<PeerListItem> {
-        let binding = self.read();
-        binding.peer_list_cache.get(mining_address).cloned()
-    }
-
-    pub fn is_a_trusted_peer(&self, miner_address: Address, source_ip: IpAddr) -> bool {
-        let binding = self.read();
-
-        let Some(peer) = binding.peer_list_cache.get(&miner_address) else {
-            return false;
-        };
-        let peer_api_ip = peer.address.api.ip();
-        let peer_gossip_ip = peer.address.gossip.ip();
-
-        let ip_matches_cached_ip = source_ip == peer_gossip_ip;
-        let ip_is_in_a_trusted_list = binding
-            .trusted_peers_api_addresses
-            .iter()
-            .any(|socket_addr| socket_addr.ip() == peer_api_ip);
-
-        ip_matches_cached_ip && ip_is_in_a_trusted_list
-    }
-
-    pub async fn request_block_from_the_network(
-        &self,
-        block_hash: BlockHash,
-        use_trusted_peers_only: bool,
-    ) -> Result<(), PeerListDataError> {
-        let (sender, receiver) = tokio::sync::oneshot::channel();
-        let message = PeerListDataMessage::RequestBlockFromNetwork {
-            block_hash,
-            use_trusted_peers_only,
-            response: sender,
-        };
-        self.0
-            .write()
-            .expect("PeerListDataInner lock poisoned")
-            .peer_list_service_sender
-            .send(message)?;
-
-        receiver.await.map_err(|recv_error| {
-            PeerListDataError::OtherInternalError(format!(
-                "Failed to receive response: {:?}",
-                recv_error
-            ))
-        })?
-    }
-
-    pub async fn request_payload_from_the_network(
-        &self,
-        evm_payload_hash: B256,
-        use_trusted_peers_only: bool,
-    ) -> Result<(), PeerListDataError> {
-        let (sender, receiver) = tokio::sync::oneshot::channel();
-        let message = PeerListDataMessage::RequestPayloadFromNetwork {
-            payload_hash: evm_payload_hash,
-            use_trusted_peers_only,
-            response: sender,
-        };
-        self.0
-            .write()
-            .expect("PeerListDataInner lock poisoned")
-            .peer_list_service_sender
-            .send(message)?;
-
-        receiver.await.map_err(|recv_error| {
-            PeerListDataError::OtherInternalError(format!(
-                "Failed to receive response: {:?}",
-                recv_error
-            ))
-        })?
-    }
-
-    pub fn read(&self) -> std::sync::RwLockReadGuard<'_, PeerListDataInner> {
-        self.0.read().expect("PeerListDataInner lock poisoned")
-    }
-
-    pub fn peer_count(&self) -> usize {
-        self.read().peer_list_cache.len()
-    }
-}
-
-#[derive(Message, Debug)]
-#[rtype(result = "()")]
-pub enum PeerListDataMessage {
-    PeerUpdated(PeerListItem),
-    RequestBlockFromNetwork {
-        block_hash: BlockHash,
-        use_trusted_peers_only: bool,
-        response: tokio::sync::oneshot::Sender<Result<(), PeerListDataError>>,
-    },
-    RequestPayloadFromNetwork {
-        payload_hash: B256,
-        use_trusted_peers_only: bool,
-        response: tokio::sync::oneshot::Sender<Result<(), PeerListDataError>>,
-    },
-}
-
-impl PeerListDataInner {
-    pub fn new(
-        config: &Config,
-        db: &DatabaseProvider,
-        peer_service_sender: UnboundedSender<PeerListDataMessage>,
-    ) -> Result<Self, PeerListDataError> {
-        let mut peer_list = Self {
-            gossip_addr_to_mining_addr_map: HashMap::new(),
-            api_addr_to_mining_addr_map: HashMap::new(),
-            peer_list_cache: HashMap::new(),
-            known_peers_cache: HashSet::new(),
-            trusted_peers_api_addresses: config
-                .node_config
-                .trusted_peers
-                .iter()
-                .map(|p| p.api)
-                .collect(),
-            peer_list_service_sender: peer_service_sender,
-        };
-
-        let read_tx = db.tx().map_err(PeerListDataError::from)?;
-
-        let peer_list_items =
-            walk_all::<PeerListItems, _>(&read_tx).map_err(PeerListDataError::from)?;
-
-        for (mining_addr, entry) in peer_list_items {
-            let address = entry.address;
-            peer_list
-                .gossip_addr_to_mining_addr_map
-                .insert(entry.address.gossip.ip(), mining_addr);
-            peer_list.peer_list_cache.insert(mining_addr, entry.0);
-            peer_list.known_peers_cache.insert(address);
-            peer_list
-                .api_addr_to_mining_addr_map
-                .insert(address.api, mining_addr);
-        }
-
-        Ok(peer_list)
-    }
-
-    pub fn add_or_update_peer(&mut self, mining_addr: Address, peer: PeerListItem) {
-        let is_updated = self.add_or_update_peer_internal(mining_addr, peer.clone());
-
-        if is_updated {
-            debug!(
-                "Sending PeerUpdated message to the service for peer {:?}",
-                mining_addr
-            );
-            // Notify the peer list service that a peer was updated
-            if let Err(e) = self
-                .peer_list_service_sender
-                .send(PeerListDataMessage::PeerUpdated(peer))
-            {
-                error!("Failed to send peer updated message: {:?}", e);
-            }
-        }
-    }
-
-    pub fn increase_score(&mut self, mining_addr: &Address, reason: ScoreIncreaseReason) {
-        if let Some(peer) = self.peer_list_cache.get_mut(mining_addr) {
-            match reason {
-                ScoreIncreaseReason::Online => {
-                    peer.reputation_score.increase();
-                }
-                ScoreIncreaseReason::ValidData => {
-                    peer.reputation_score.increase();
-                }
-            }
-        }
-    }
-
-    pub fn decrease_peer_score(&mut self, mining_addr: &Address, reason: ScoreDecreaseReason) {
-        if let Some(peer_item) = self.peer_list_cache.get_mut(mining_addr) {
-            match reason {
-                ScoreDecreaseReason::BogusData => {
-                    peer_item.reputation_score.decrease_bogus_data();
-                }
-                ScoreDecreaseReason::Offline => {
-                    peer_item.reputation_score.decrease_offline();
-                }
-            }
-
-            // Don't propagate inactive peers
-            if !peer_item.reputation_score.is_active() {
-                self.known_peers_cache.remove(&peer_item.address);
-            }
-        }
-    }
-
-    /// Add a peer to the peer list. Returns true if the peer was added, false if it already exists.
-    fn add_or_update_peer_internal(&mut self, mining_addr: Address, peer: PeerListItem) -> bool {
-        let gossip_addr = peer.address.gossip;
-        let peer_address = peer.address;
-
-        if let std::collections::hash_map::Entry::Vacant(peer_cache) =
-            self.peer_list_cache.entry(mining_addr)
-        {
-            debug!("Adding peer {:?} to the peer list", mining_addr);
-            peer_cache.insert(peer);
-            self.gossip_addr_to_mining_addr_map
-                .insert(gossip_addr.ip(), mining_addr);
-            self.api_addr_to_mining_addr_map
-                .insert(peer_address.api, mining_addr);
-            self.known_peers_cache.insert(peer_address);
-            debug!(
-                "Peer {:?} added to the peer list with address {:?}",
-                mining_addr, peer_address
-            );
-            true
-        } else {
-            debug!(
-                "Peer {:?} already exists in the peer list, checking if the address needs updating",
-                mining_addr
-            );
-            if let Some(existing_peer) = self.peer_list_cache.get_mut(&mining_addr) {
-                let handshake_cooldown_expired =
-                    existing_peer.last_seen + HANDSHAKE_COOLDOWN < peer.last_seen;
-                existing_peer.last_seen = peer.last_seen;
-                if existing_peer.address != peer_address {
-                    debug!("Peer address mismatch, updating to new address");
-                    self.update_peer_address(mining_addr, peer_address);
-                    true
-                } else if handshake_cooldown_expired {
-                    debug!("Peer address is the same, but the handshake cooldown has expired, so we need to re-handshake");
-                    true
-                } else {
-                    debug!("Peer address is the same, no update needed");
-                    false
-                }
-            } else {
-                warn!(
-                    "Peer {:?} is not found in the peer list cache, which shouldn't happen",
-                    mining_addr
-                );
-                false
-            }
-        }
-    }
-
-    fn update_peer_address(&mut self, mining_addr: Address, new_address: PeerAddress) {
-        if let Some(peer) = self.peer_list_cache.get_mut(&mining_addr) {
-            let old_address = peer.address;
-            peer.address = new_address;
-            self.gossip_addr_to_mining_addr_map
-                .remove(&old_address.gossip.ip());
-            self.gossip_addr_to_mining_addr_map
-                .insert(new_address.gossip.ip(), mining_addr);
-            self.known_peers_cache.remove(&old_address);
-            self.known_peers_cache.insert(old_address);
-            self.api_addr_to_mining_addr_map.remove(&old_address.api);
-            self.api_addr_to_mining_addr_map
-                .insert(new_address.api, mining_addr);
         }
     }
 }
@@ -702,20 +288,8 @@ where
 
         ctx.run_interval(INACTIVE_PEERS_HEALTH_CHECK_INTERVAL, |act, ctx| {
             // Collect inactive peers with the required fields
-            let inactive_peers: Vec<(Address, PeerListItem, SocketAddr)> = act
-                .peer_list_data_guard
-                .read()
-                .peer_list_cache
-                .iter()
-                .filter(|(_mining_addr, peer)| !peer.reputation_score.is_active())
-                .map(|(mining_addr, peer)| {
-                    // Clone or copy the fields we need for the async operation
-                    let peer_item = peer.clone();
-                    let mining_addr = *mining_addr;
-                    let peer_addr = peer_item.address;
-                    (mining_addr, peer_item, peer_addr.gossip)
-                })
-                .collect();
+            let inactive_peers: Vec<(Address, PeerListItem, SocketAddr)> =
+                act.peer_list_data_guard.inactive_peers();
 
             for (mining_addr, peer, ..) in inactive_peers {
                 // Clone the peer address to use in the async block
@@ -744,10 +318,7 @@ where
         // Initiate the trusted peers handshake
         let trusted_peers_handshake_task = Self::trusted_peers_handshake_task(
             peer_service_address.clone(),
-            self.peer_list_data_guard
-                .read()
-                .trusted_peers_api_addresses
-                .clone(),
+            self.peer_list_data_guard.trusted_peer_addresses(),
         )
         .into_actor(self);
         ctx.spawn(trusted_peers_handshake_task);
@@ -755,7 +326,7 @@ where
         // Announce yourself to the network
         let version_request = self.create_version_request();
         let api_client = self.irys_api_client.clone();
-        let peers_cache = self.peer_list_data_guard.read().known_peers_cache.clone();
+        let peers_cache = self.peer_list_data_guard.all_known_peers();
         let announce_fut = Self::announce_yourself_to_all_peers(
             api_client,
             version_request,
@@ -798,34 +369,6 @@ pub enum PeerListServiceError {
     FailedToRequestData(String),
 }
 
-#[derive(Debug, Error)]
-pub enum PeerListDataError {
-    #[error("Internal database error: {0}")]
-    Database(DatabaseError),
-    #[error("Peer list internal error: {0:?}")]
-    InternalSendError(SendError<PeerListDataMessage>),
-    #[error("Peer list internal error: {0}")]
-    OtherInternalError(String),
-}
-
-impl From<SendError<PeerListDataMessage>> for PeerListDataError {
-    fn from(err: SendError<PeerListDataMessage>) -> Self {
-        Self::InternalSendError(err)
-    }
-}
-
-impl From<DatabaseError> for PeerListDataError {
-    fn from(err: DatabaseError) -> Self {
-        Self::Database(err)
-    }
-}
-
-impl From<eyre::Report> for PeerListDataError {
-    fn from(err: eyre::Report) -> Self {
-        Self::Database(DatabaseError::Other(err.to_string()))
-    }
-}
-
 impl From<MailboxError> for PeerListServiceError {
     fn from(value: MailboxError) -> Self {
         Self::InternalSendError(value)
@@ -844,18 +387,6 @@ impl From<eyre::Report> for PeerListServiceError {
     }
 }
 
-#[derive(Clone, Debug, Copy)]
-pub enum ScoreDecreaseReason {
-    BogusData,
-    Offline,
-}
-
-#[derive(Clone, Debug, Copy)]
-pub enum ScoreIncreaseReason {
-    Online,
-    ValidData,
-}
-
 impl<A, R> PeerListServiceWithClient<A, R>
 where
     A: ApiClient,
@@ -864,7 +395,11 @@ where
     fn flush(&self) -> Result<(), PeerListServiceError> {
         if let Some(db) = &self.db {
             db.update(|tx| {
-                for (addr, peer) in self.peer_list_data_guard.read().peer_list_cache.iter() {
+                for (addr, peer) in self
+                    .peer_list_data_guard
+                    .all_know_peers_with_mining_address()
+                    .iter()
+                {
                     insert_peer_list_item(tx, addr, peer).map_err(PeerListServiceError::from)?;
                 }
                 Ok(())
@@ -1016,7 +551,7 @@ where
     async fn announce_yourself_to_all_peers(
         api_client: A,
         version_request: VersionRequest,
-        known_peers_cache: HashSet<PeerAddress>,
+        known_peers_cache: Vec<PeerAddress>,
         peer_service_address: Addr<Self>,
         reth_service_address: Option<Addr<R>>,
     ) {
@@ -1182,9 +717,7 @@ where
 
         let already_in_cache = self
             .peer_list_data_guard
-            .read()
-            .api_addr_to_mining_addr_map
-            .contains_key(&msg.api_address);
+            .contains_api_address(&msg.api_address);
         let already_announcing = self
             .currently_running_announcements
             .contains(&msg.api_address);
@@ -1383,6 +916,8 @@ where
 mod tests {
     use super::*;
     use irys_api_client::test_utils::CountingMockClient;
+    use irys_database::tables::PeerListItems;
+    use irys_database::walk_all;
     use irys_storage::irys_consensus_data_db::open_or_create_irys_consensus_data_db;
     use irys_testing_utils::utils::setup_tracing_and_temp_dir;
     use irys_types::peer_list::PeerScore;
@@ -1924,7 +1459,7 @@ mod tests {
             true,
             None,
         );
-        let known_peers: HashSet<_> = vec![peer1.address, peer2.address].into_iter().collect();
+        let known_peers = vec![peer1.address, peer2.address];
         let version_request = VersionRequest::default();
 
         PeerListServiceWithClient::announce_yourself_to_all_peers(
